separated collider management from state management.
